+++ paddle/cinn/api/op_topo_pattern.h
@@ -1,7 +1,22 @@
+// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #pragma once
 
-#include <vector>
 #include <list>
+#include <variant>
+#include <vector>
 
 namespace cinn::api {
 
@@ -16,16 +31,17 @@ struct InjectiveSourcePattern {};
 template <typename T>
 struct SingleReductionOpPattern {};
 
-// ElementWise/Broadcast ops which have shardable dimentions and reduction ancestors.
+// ElementWise/Broadcast ops which have shardable dimentions and reduction
+// ancestors.
 template <typename T>
 struct PartialShardablePattern {};
 
 // Reduce base pattern
 template <typename T>
 struct ReductionPattern {
-
   using Nothing = std::monostate;
-  std::variant<Nothing, InjectiveSourcePattern<T>, PartialShardablePattern<T>> input;
+  std::variant<Nothing, InjectiveSourcePattern<T>, PartialShardablePattern<T>>
+      input;
   SingleReductionOpPattern<T> reduce_op_pattern;
 
   bool HasFusedInput() const {
@@ -36,7 +52,9 @@ struct ReductionPattern {
 // Stmt := IS | R | PS
 // ops in StmtPattern will be lowered into a inlined cuda code.
 template <typename T>
-using StmtPattern = std::variant<InjectiveSourcePattern<T>, ReductionPattern<T>, PartialShardablePattern<T>>;
+using StmtPattern = std::variant<InjectiveSourcePattern<T>,
+                                 ReductionPattern<T>,
+                                 PartialShardablePattern<T>>;
 
 // Stmts := [Stmt]
 template <typename T>
@@ -56,4 +74,4 @@ using StmtPatternVec = std::vector<StmtPattern<T>>;
 template <typename T>
 using OpTopoPattern = std::variant<ErrorPattern<T>, StmtPatternVec<T>>;
 
-}
+}  // namespace cinn::api
